requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
sqlalchemy
python-dotenv
tensorflow
scikit-learn
kagglehub
alembic
pytest